refactor: a lot of cleanup
