+++ packages/shared/src/components/profile/devcard/index.ts
@@ -1,1 +1,6 @@
 export * from './DevCard';
+export * from './DevCardStats';
+export * from './DevCardStatsSection';
+export * from './DevCardFooter';
+export * from './DevCardContainer';
+export * from './common';
